Convert distance in liters of water
A function was created to convert the distance measured with an ultrasonic sensor to liters of water in the tank. To make this function I did a linear progression with a relationship between liters of water and the installation distance ultrasonic sensor. Finally, it shows this measure through the TM1637 display.